+++ app/filters/dorsale/customer_vault/small_data/filter_strategy_by_activity_type.rb
@@ -1,0 +1,7 @@
+class Dorsale::CustomerVault::SmallData::FilterStrategyByActivityType < ::Agilibox::SmallData::FilterStrategy
+  def apply(query, value)
+    corporations_id = query.where(activity_type_id: value).pluck(:id)
+    individuals_id  = query.where(corporation_id: corporations_id).pluck(:id)
+    query.where(id: (corporations_id + individuals_id))
+  end
+end